+++ lib/src/storage.dart
@@ -11,6 +11,7 @@ abstract class StorageService {
   Future<void> store(String key, String value);
   Future<String> retrieve(String key);
   Future<void> delete(String key);
+  Future<bool> contains(String key);
 }
 
 class _SecureStorageService implements StorageService {
@@ -36,5 +37,11 @@ class _SecureStorageService implements StorageService {
     given(key, "key").ensureHasValue();
 
     await this._secureStorage.delete(key: key.trim());
+  }
+
+  Future<bool> contains(String key) async {
+    given(key, "key").ensureHasValue();
+
+    return this._secureStorage.containsKey(key: key);
   }
 }